Release v2.0 - Mass cleanup and new features

+++ events/manager/queueEnd.js
@@ -2,6 +2,11 @@ const { MessageEmbed } = require('discord.js');
 
 module.exports = async (client, player) => {
     let channel = client.channels.cache.get(player.textChannel);
+    if (player.n === true) {
+        client.update(channel.guild.id, true);
+        player.destroy();
+        return;
+    }
     if (channel.lastNowPlayingMessage) channel.lastNowPlayingMessage.delete();
     channel.lastNowPlayingMessage = false;
     const embed = new MessageEmbed()
